--- frontend/src/app/(routes)/authz/components/GrantAuthzForms.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const GrantAuthzForms = () => {
-  return <div>GrantAuthzForms</div>;
-};
-
-export default GrantAuthzForms;
